requirements.txt
aiocqhttp==1.3.0
opencc_python_reimplemented==0.1.6
APScheduler==3.6.3
requests==2.25.1
numpy==1.19.3
retrying==1.3.3
Quart==0.11.5
six==1.15.0
Jinja2==2.11.3
expiringdict==1.2.1
httpx==0.17.1
tzlocal==2.1
aiohttp==3.7.3
arrow==0.17.0
feedparser==6.0.2
opencv_python==4.5.2.52
peewee==3.14.4
aiohttp_proxy==0.1.2
ujson==4.0.2
lxml==4.6.2
tweepy==3.10.0
nonebot==1.8.3
loguru==0.5.3
beautifulsoup4==4.9.3
dataclasses==0.8
opencc==1.1.2
Pillow==8.2.0
